Legacy code
I've invested over 190 hours into this project (~8 days), not counting breaks over a few months, but its kinda spahgetti so i'm restarting. also i'm going to add an even older version's source code (From ‎May ‎13 of this year)  to this repo. also, just checked, the version after that was created ‎May ‎25, so that's when i started working on this.