Preparing rewrite
Instead of trying to workaround the existing GUI code, i'm going to rewrite it completely. Since this is just a gui and not a complete program rewrite it should be easy to plug back in the conversion code, and this rewrite will bring a new GUI design, which will have more options and be more intuitive. when i started this project i didn't think i would be exporting to OBJ, or needing to cull blocks, or include bushy leaves, etc, so this version will have options for those things.